requirements.txt
google-generativeai
pyyaml
schedule